+++ lib/src/infra/models/pagination_model.dart
@@ -3,19 +3,19 @@ import '../../domain/enums/sorting_order_type_enum.dart';
 
 class PaginationModel extends PaginationEntity {
   PaginationModel({
-    super.pageNumber,
     super.limit,
     super.offset,
     super.sortingBy,
+    super.pageNumber,
     super.sortingOrder,
   });
 
   factory PaginationModel.fromMap(Map<String, dynamic> map) {
     return PaginationModel(
-      pageNumber: map['page']?.toInt() ?? 1,
-      offset: map['offset']?.toInt() ?? 10,
-      limit: map['limit']?.toInt() ?? 10,
       sortingBy: map['sortingBy'] ?? '',
+      limit: int.tryParse(map['limit'].toString()),
+      offset: int.tryParse(map['offset'].toString()),
+      pageNumber: int.tryParse(map['page'].toString()),
       sortingOrder: sortingOrderTypeFromJson(map['sortingOrder']),
     );
   }
@@ -23,8 +23,8 @@ class PaginationModel extends PaginationEntity {
   factory PaginationModel.fromEntity(PaginationEntity entity) {
     return PaginationModel(
       sortingOrder: entity.sortingOrder,
-      sortingBy: entity.sortingBy,
       pageNumber: entity.pageNumber,
+      sortingBy: entity.sortingBy,
       offset: entity.offset,
       limit: entity.limit,
     );
@@ -32,11 +32,11 @@ class PaginationModel extends PaginationEntity {
 
   Map<String, dynamic> get toMap {
     return {
-      'page': pageNumber,
-      'offset': offset,
-      'limit': limit,
-      'sortingOrder': sortingOrder.toJson,
+      if (limit != null) 'limit': limit,
+      if (offset != null) 'offset': offset,
+      if (pageNumber != null) 'page': pageNumber,
       if (sortingBy.isNotEmpty) 'sortingBy': sortingBy,
+      if (sortingOrder != null) 'sortingOrder': sortingOrder!.toJson,
     };
   }
 }