+++ lib/src/infra/models/received_notifications_model.dart
@@ -15,10 +15,10 @@ class ReceivedNotificationModel extends ReceivedNotificationEntity
 
   factory ReceivedNotificationModel.fromMap(Map<String, dynamic> map) {
     return ReceivedNotificationModel(
-      id: map['id']?.toInt() ?? 0,
       title: map['title'],
       body: map['body'],
       payload: map['payload'],
+      id: int.tryParse(map['id'].toString()) ?? 0,
     );
   }