+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.4
+
+* Fix token authorization and api key into http client
+* Update pagination params
+
+## 1.0.3
+
+* Improve Local auth driver
+
 ## 1.0.2
 
 * Update EventLog